rewrote code to collect data
I think it's not quite accurate, but unsure. I think we may need to create a new object (or reset everything like in constructor) every time we reset the value because otherwise too much data is similar. Unsure.
There used to be a bug when I just reset specific values, which went away when I replaced the object in between different parameters tested. I think this means my reset values functions need to be more complicated than normal setters. Will check out in the morning.
Code is currently running on desktop in 224 but I wouldn't trust the data.