This is another way I wrote the code to collect data
It's currently running in 224 instead of the old one because I knew the data was bad.
I have a feeling this will work better since every time the value changes, the entire code is run again anyway, so it's better to reset to initial free flow conditions, etc. Otherwise we're doing math on the wrong stuff/ an already solved problem that we screwed up a variable of, which seems like the wrong idea.